Created a util tool to add coverage instrumentation to go binaries
This commit adds the `instrumentmain.go` file, which is a utility script which
can instrument regular go source files to be venerable to coverage analysis,
just like that done by `go test -cover`. The difference however, is that this
does not employ any of the scaffolding that comes with the `go test` framework.
Instead it leverages the `go tool cover` functionality, to mimick the same
functionality, only that the binary runs just like it normally would. This also
means that the source code can be built just like it normally would.

In short, `go tool cover` adds a struct - GoCover - to each file it is told to
instrument. Hence, this tool runs `go tool cover` on all files in the packages
listed on the command line. Each `GoCover` struct in each file, is given an
unique suffix, and the names collected in an internal structure. After all the
filesa have been instrumented in this way, the variable names, and the package
names that have been collected are then used to dynamically generate a new
`main.go` file on stdout. This generation is the only way, for go source code to
reach all the generated `GoCover<suffix>` structs that have been generated by
the coverage tool.

The generated `main.go` should function just like the regular `main.go` file in
all regards, except for that it in the end collects the coverage before returning.

For more information, see the usage information in the binary.

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>